ZCII-3396: Les utilisateurs ayant le GDSC dans leur espace personnel peuvent supprimer des sites officiels

+++ osyl-common/osyl-common-impl/impl/src/java/org/sakaiquebec/opensyllabus/common/impl/OsylSiteServiceImpl.java
@@ -2165,7 +2165,11 @@ public class OsylSiteServiceImpl implements OsylSiteService, EntityTransferrer {
 	try {
 	    securityService.pushAdvisor(advisor);
 	    site = getSite(siteId);
-	    siteService.removeSite(site);
+	    
+	    //Do not allow sites with providerId to be deleted
+	    if (site.getProviderGroupId() == null){
+	    	siteService.removeSite(site);
+	    }
 	} catch (IdUnusedException e) {
 	    log.info("User " + sessionManager.getCurrentSession().getUserEid()
 		    + " can not delete the site " + siteId
